Move value_table
