+++ packages/server/computation_container/test/integration_test/join_table_test.hpp
@@ -3,8 +3,8 @@
 #include <vector>
 
 #include "client/computation_to_db/client.hpp"
-#include "client/computation_to_db/join_table.hpp"
 #include "gtest/gtest.h"
+#include "job/table_data/join_table.hpp"
 namespace fs = std::experimental::filesystem;
 
 using Schema = std::vector<qmpc::ComputationToDb::SchemaType>;
